# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The Road Of Code",
        "title": "Road",
        "contact": {},
        "license": {
            "name": "Raja",
            "url": "https://github.com/0RAJA"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/comment": {
            "get": {
                "description": "根据post_id和偏移量显示一个帖子的部分评论的内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "显示评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "需要显示评论的帖子ID",
                        "name": "PostID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码 default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量 default and max 10",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回评论的信息",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/logic.ListCommentByPostIDReply"
                            }
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "修改一个评论的内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "修改评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "需要修改的评论ID",
                        "name": "comment_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "评论修改后的内容 1\u003c=len\u003c=100",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个对于帖子的评论或对评论的回复",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "创建评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "帖子ID",
                        "name": "post_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "用户名",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "评论内容 1\u003c=len\u003c=100",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "回复的评论ID",
                        "name": "to_comment_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/comment/{comment_id}": {
            "delete": {
                "description": "删除一个评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "需要删除的评论ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errcode.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "logic.Comment": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "description": "头像链接",
                    "type": "string"
                },
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "create_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "depository_url": {
                    "description": "仓库连接",
                    "type": "string"
                },
                "id": {
                    "description": "评论ID",
                    "type": "integer"
                },
                "modify_time": {
                    "description": "修改时间",
                    "type": "string"
                },
                "post_id": {
                    "description": "帖子ID",
                    "type": "integer"
                },
                "to_comment_id": {
                    "description": "回复的评论的ID",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "logic.ListCommentByPostIDReply": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/logic.Comment"
                    }
                },
                "pager": {
                    "$ref": "#/definitions/logic.Pager"
                }
            }
        },
...
        "logic.Pager": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "页数",
                    "type": "integer"
                },
                "page_size": {
                    "description": "一页的大小",
                    "type": "integer"
                },
                "total_rows": {
                    "description": "结果的总数量",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}