# internal/dao/mysql/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.

package db

import (
	"time"
)

type Comment struct {
	ID          int64     `json:"id"`
	PostID      int64     `json:"post_id"`
	Username    string    `json:"username"`
	Content     string    `json:"content"`
	ToCommentID int64     `json:"to_comment_id"`
	CreateTime  time.Time `json:"create_time"`
	ModifyTime  time.Time `json:"modify_time"`
}

type Manager struct {
	Username  string `json:"username"`
	Password  string `json:"password"`
	AvatarUrl string `json:"avatar_url"`
}

type Post struct {
	ID         int64     `json:"id"`
	Cover      string    `json:"cover"`
	Title      string    `json:"title"`
	Abstract   string    `json:"abstract"`
	Content    string    `json:"content"`
	Public     bool      `json:"public"`
	Deleted    bool      `json:"deleted"`
	CreateTime time.Time `json:"create_time"`
	ModifyTime time.Time `json:"modify_time"`
}

type PostNum struct {
	PostID     int64 `json:"post_id"`
	StarNum    int64 `json:"star_num"`
	VisitedNum int64 `json:"visited_num"`
}

type PostTag struct {
	ID     int64 `json:"id"`
	PostID int64 `json:"post_id"`
	TagID  int64 `json:"tag_id"`
}

type Tag struct {
	ID         int64     `json:"id"`
	TagName    string    `json:"tag_name"`
	CreateTime time.Time `json:"create_time"`
}

type Top struct {
	ID     int64 `json:"id"`
	PostID int64 `json:"post_id"`
}

type User struct {
	Username      string    `json:"username"`
	AvatarUrl     string    `json:"avatar_url"`
	DepositoryUrl string    `json:"depository_url"`
	Address       string    `json:"address"`
	CreateTime    time.Time `json:"create_time"`
	ModifyTime    time.Time `json:"modify_time"`
}

type UserStar struct {
	ID       int32  `json:"id"`
	Username string `json:"username"`
	PostID   int64  `json:"post_id"`
}

type View struct {
	ID         int64     `json:"id"`
	ViewsNum   int64     `json:"views_num"`
	CreateTime time.Time `json:"create_time"`
}